Moved constants and used it for hardcoded paths
