+++ src/cli/commands/help.js
@@ -2,7 +2,7 @@ import chalk from 'chalk'
 import { forEach, map, padEnd } from 'lodash'
 
 import commands from '.'
-import * as constants from '../constants'
+import * as constants from '../../constants'
 import * as utils from '../utils'
 
 export const usage = 'help [command]'
