+++ src/constants.js
@@ -1,0 +1,6 @@
+import path from 'path'
+
+export const cli = 'tailwind'
+export const defaultConfigFile = './tailwind.config.js'
+export const defaultConfigStubFile = path.resolve(__dirname, '../stubs/defaultConfig.stub.js')
+export const simpleConfigStubFile = path.resolve(__dirname, '../stubs/simpleConfig.stub.js')
