+++ src/index.js
@@ -8,6 +8,7 @@ import perfectionist from 'perfectionist'
 import registerConfigAsDependency from './lib/registerConfigAsDependency'
 import processTailwindFeatures from './processTailwindFeatures'
 import resolveConfig from './util/resolveConfig'
+import { defaultConfigFile } from './constants'
 
 import defaultConfig from '../stubs/defaultConfig.stub.js'
 
@@ -21,7 +22,7 @@ function resolveConfigPath(filePath) {
   }
 
   try {
-    const defaultConfigPath = path.resolve('./tailwind.config.js')
+    const defaultConfigPath = path.resolve(defaultConfigFile)
     fs.accessSync(defaultConfigPath)
     return defaultConfigPath
   } catch (err) {
